Two fixes for EUST, v2.9.2
This SHOULD run faster (and so hang less) when updating the save/unsave
buttons across all pages. Instead of iterating over every single saved
tag, it only iterates over the actual save/unsave links. In a future
update, I might make it compare the length of the lists to try and pick
the shorter one.