Improved san checks in corrupt.browser.js
Add san check to avoid operating on nonexistent portions of the
current/target strings.

Add san check to prevent endless loop when operable portions of the
current/target strings are the same, even if the rest isn't.

Should now be guaranteed safe to use on origin and target strings of
differing lengths in either order, without internally attempting to
compare and modify nonexistent characters for corruption.