Enhanced621 v1.6.0 - double update
Post pages with a "current search" bar at the top (site feature) will
now link to a search for the searched tags. For instance, if you're
viewing a post from a search for your favourite artist, you can now
click the search string in the navbar above the image to go back to that
search page.

When viewing a post from a search, whatever tags you searched for that
are present on that post will be elevated to the top of their tag group
and italicised. This does not include metadata tags (including the
`rating:*` one) because they don't show up in the tag list, and it
doesn't include wildcard tags (such as, from the site's search help,
`african_*`) for technical reasons.