en621 v2.1.0: pool page direct image link toggle
On pool pages, there is now a toggle in the lower left corner to switch
between linking to post pages and linking directly to the full-size
image. This applies both in and out of pool reader mode, and when in
reader mode will affect both images that have been loaded and images
that are being loaded.

This MAY be extended to post index pages (tag searches) at some point,
but it was felt that generally users want to see the post page when
browsing by tags.