en621 v2.1.1: bugfix post rating in sidebar
Fixed a bug where the segment that puts the post rating in the sidebar
would throw (and be caught, but it wouldn't FINISH) when there was no
existing search parameter on a post page.

Also noticed that if you don't have an existing search on a post page,
the +/- links by each tag in the listing aren't shown. That feels like
an intended decision, but a bad one. There'll probably be a feature to
put them in anyway at some point.