en621 v2.2.0: scope expansion for three features
Tag elevation/highlighting (italicisation) extends to post index/search
pages now, as does the direct image link toggle. The alt-q keybind to
focus the search box (given in the tooltip text for it) has been
implemented, since it's not a functional part of the site itself.

A planned features list has been added in the header comments.