en621 v2.5.1: pool reader stability
Pool reader mode no longer stalls out and dies when encountering a
deleted post or a video. There is not yet a video player, since that
would require redesigning the post links, but the preview image will be
shown along with a large red tag indicating that it is a preview only.
This tag will disappear when the cursor is over the preview image.

Additionally, some minor bugs were fixed along the way. Maybe that's the
real treasure here.