en621 v2.9.0: reduced network load
Pool reader mode will still automatically activate if the URL hash is
`#reader-mode` to allow page reloads to automatically apply it, but
links to pool pages will no longer automatically be edited to make this
happen by default. This should reduce network load, both for clients and
for the site itself.

Please be advised that automatic reader mode is being considered for
removal. When using a tab suspender addon or restoring an old browsing
session after restarting the browser, pages will (currently) return to
the mode they were last in, which may put extra stress on the network at
a time when many pages might also be loading, in the case of restored
sessions. It is possible that automatic pool reader activation will be
removed in a future update.