HVEU v3.2.1: bugfix an edge case
Whenever the last event was at least 24 hours ago, the auto-reloader
triggers and reloads the page. But if there was no event on page load,
then the timestamp of the last event never got updated, which left the
script in an infinite reload loop. Now, the auto-reloader also updates
the last event to be the time that it reloaded the page.