en621 v3.0.0: event and API update
In the interests of allowing other scripts to interact with en621, the
script now exposes the main parts of its callable functionality on the
page via the (read-only) `EN621_API` object. This does not include
run-once features, only things like controlling pool reader mode,
putting user messages on the screen, registering keybinds for its own
handler, and accessing the current search query.

Additionally, events are now dispatched to the `document` for certain
occurances. There's a built in event debugger display if you want to see
what happens while you play around with it. At present, functionality
exists to allow cancelable events but none of them presently are - the
cancelability of an en621 event is also displayed by the debug handler.

Please note that the helper function that en621 defines to send its
events to the page is NOT exposed via the API - this is an intentional
decision. The `en621` event name should be considered reserved for
events coming from the script itself, and any addon scripts should
choose their own event name(s) to use.

Finally, a minor bug was fixed where non-gallery pages under the
`/gallery/` namespace would still be treated as galleries. The check is
slightly more stringent now.

Of interest only to people reading the source, the changelog has been
reversed so that the most recent updates are at the top, and major
versions are now separated by a single blank line. Some minor cleanup
has also been added in a few places.