en621 v3.1.0: this should've been in the last one
Message boxes now use the right offset. Messages can now have an
automatic timeout. You can still pass a string (of HTML) to use as a
message, or you can now pass an element directly, or an array of them.
Messages closing (or timing out) now post an event, which includes the
cause. The `has-related-posts` flag is now only set when there actually
are related posts. There's a new `post-in-pool` flag. Pages autoscroll
to the actual post content, and as a result there are now notices when
the post has related posts or is in a pool.