en621 v3.5.0: control/status tabs
The direct image link toggle was previously a one-off, but is now part
of the new control/status tabs functionality. This is exposed via the
API in the form of `addControlTab()`. Furthermore, this new feature is
already in use by pool reader as well, providing a small status readout
of how many images in a pool have been loaded, so you can read while it
loads and still have an easy view of the progress so far.

Yes, I know the API isn't actually DOCUMENTED anywhere yet, I'll get to
that soon. Probably. The ideas are coming fast right now. I've only had
a break to sleep since the last batch.