en621 v4.1.0: basic version checking
A label is now added to the navbar showing the currently loaded version
of en621. Clicking this label will do an inter-tab version comparison,
and if any tabs are on different version, all tabs with older versions
will reload to update to match the latest version. What this means is
that if you have a bunch of e621 tabs open, update the script, and then
open a new tab, you can click the label to tell all tabs on an older
version to update. Due to the mechanics by which the communication and
comparison are implemented, you can even click the label in one of the
older tabs - the ONLY requirement is that a tab running a later version
be open SOMEWHERE.

Conveniently, if you have tab A running two version behind and tab B
running one version behind, using the updater will update tab A to the
latest version, and then using the updater again - even from the same
tab - will update tab B.

Please note that the updater works by RELOADING the tabs on an older
version, and also that it does not check the remote hosted version; it
is technically only a version comparison between all open tabs running
en621 v4.1.0 or later. Tampermonkey's built-in update check already
handles remote script update checking, this just lets you bump all your
open tabs to the latest version when you DO update.