en621 documentation link
The bottom of the file, where the API object is defined, now has a link
to the documentation. I also took out an unnecessary `else` block while
I was at it. Added some more possible ideas to the plans section too.

This is NOT an actual script update. Nothing has changed regarding
functionality, internally or externally. Users will not be prompted to
update, because the version isn't changing.