en621 v4.3.0: tooltip controls
When using Hover Zoom (or a similar extension) to view full-size images
without loading their entire page, the tooltips containing image tags
and metadata can get in the way. This update provides a feature to
toggle those tooltips on and off, with both a control tab and a hotkey.

If hover zoom is detected on the page at load time, tooltips will be
automatically disabled. This may be subject to load order and timing,
so if it fails (particularly consistently) please open an issue to let
me know and I'll look into it.