en621 v4.4.0: last-seen post tracking
Only applies to post index (search) pages. The most recent post on a
search page will be remembered (mapped to that search) and will be
highlighted on subsequent searches with the same query, to make it
easier to see what's new since the last search.

Includes a new flag. Flags now also emit events when set or unset.