en621 v4.5.0: `hasNewPosts()` API function
There are no user-facing changes in this update. The API object has been cleaned (no longer has an object prototype) and has a new function to check if the current page is a post index page that has unseen posts.

A framework now exists for broadcasting all-tabs commands to execute, but there are no actual commands implemented for it yet. Consider it a plugin point for future expansion.