en621 v6.0.0: cancelable message tabs
It is now possible to allow canceling the creation and removal of message tabs. When calling `putMessage()` directly, the default is to forbid event cancelation, but a new parameter exists to set the dispatched event to cancelable.

The interface for `putError()` and `putHelp()` have not changed; errors are non-cancelable while help messages may be canceled. `putWarning()` defaults to non-cancelable, but includes a new parameter to allow cancelation.

The only potentially-breaking change is that events are now necessarily fired *before* the DOM is changed. If previous event handlers relied on the message tab existing for some reason, they will need to be updated.